Moved get functionality to individual file and got rid of central file
